move ImageUploader to shard folder
