+++ app/javascript/components/shard/singleImageUploader.vue
